+++ heron/packing/src/java/com/twitter/heron/packing/NullPacking.java
@@ -17,8 +17,8 @@ package com.twitter.heron.packing;
 import java.util.HashMap;
 
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.PackingPlan;
 import com.twitter.heron.spi.packing.IPacking;
+import com.twitter.heron.spi.packing.PackingPlan;
 
 public class NullPacking implements IPacking {
