+++ heron/scheduler-core/src/java/com/twitter/heron/scheduler/server/RestartRequestHandler.java
@@ -21,8 +21,8 @@ import com.sun.net.httpserver.HttpHandler;
 
 import com.twitter.heron.proto.scheduler.Scheduler;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.utils.NetworkUtils;
 import com.twitter.heron.spi.scheduler.IScheduler;
+import com.twitter.heron.spi.utils.NetworkUtils;
 import com.twitter.heron.spi.utils.SchedulerUtils;
 
 public class RestartRequestHandler implements HttpHandler {
