+++ heron/scheduler-core/src/java/com/twitter/heron/scheduler/server/SchedulerServer.java
@@ -23,8 +23,8 @@ import java.util.concurrent.Executors;
 import com.sun.net.httpserver.HttpServer;
 
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.utils.NetworkUtils;
 import com.twitter.heron.spi.scheduler.IScheduler;
+import com.twitter.heron.spi.utils.NetworkUtils;
 
 public class SchedulerServer {
