+++ heron/scheduler-core/tests/java/com/twitter/heron/scheduler/server/SchedulerServerTest.java
@@ -20,8 +20,8 @@ import org.mockito.Mockito;
 
 import com.twitter.heron.common.basics.SysUtils;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.utils.NetworkUtils;
 import com.twitter.heron.spi.scheduler.IScheduler;
+import com.twitter.heron.spi.utils.NetworkUtils;
 
 public class SchedulerServerTest {
   @Test
