+++ heron/schedulers/src/java/com/twitter/heron/scheduler/NullLauncher.java
@@ -15,7 +15,7 @@
 package com.twitter.heron.scheduler;
 
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.PackingPlan;
+import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.ILauncher;
 
 public class NullLauncher implements ILauncher {
