+++ heron/schedulers/src/java/com/twitter/heron/scheduler/NullScheduler.java
@@ -19,7 +19,7 @@ import java.util.List;
 
 import com.twitter.heron.proto.scheduler.Scheduler;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.PackingPlan;
+import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.IScheduler;
 
 public class NullScheduler implements IScheduler {
