+++ heron/schedulers/src/java/com/twitter/heron/scheduler/local/LocalLauncher.java
@@ -25,11 +25,11 @@ import java.util.logging.Logger;
 import com.twitter.heron.common.basics.SysUtils;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.Context;
-import com.twitter.heron.spi.common.PackingPlan;
-import com.twitter.heron.spi.utils.ShellUtils;
+import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.ILauncher;
 import com.twitter.heron.spi.utils.Runtime;
 import com.twitter.heron.spi.utils.SchedulerUtils;
+import com.twitter.heron.spi.utils.ShellUtils;
 
 /**
  * Launch topology locally to a working directory.
