+++ heron/schedulers/src/java/com/twitter/heron/scheduler/slurm/SlurmLauncher.java
@@ -20,7 +20,7 @@ import java.util.logging.Logger;
 
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.Context;
-import com.twitter.heron.spi.common.PackingPlan;
+import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.ILauncher;
 import com.twitter.heron.spi.utils.Runtime;
 import com.twitter.heron.spi.utils.SchedulerUtils;
