+++ heron/schedulers/tests/java/com/twitter/heron/scheduler/local/LocalLauncherTest.java
@@ -27,7 +27,7 @@ import org.powermock.modules.junit4.PowerMockRunner;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.ConfigKeys;
 import com.twitter.heron.spi.common.Keys;
-import com.twitter.heron.spi.common.PackingPlan;
+import com.twitter.heron.spi.packing.PackingPlan;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(LocalContext.class)
