+++ heron/spi/src/java/com/twitter/heron/spi/packing/IPacking.java
@@ -15,7 +15,6 @@
 package com.twitter.heron.spi.packing;
 
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.PackingPlan;
 
 /**
  * Packing algorithm to use for packing multiple instances into containers. Packing hints like
