+++ heron/spi/src/java/com/twitter/heron/spi/scheduler/IScheduler.java
@@ -19,7 +19,7 @@ import java.util.List;
 
 import com.twitter.heron.proto.scheduler.Scheduler;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.PackingPlan;
+import com.twitter.heron.spi.packing.PackingPlan;
 
 /**
  * Scheduler object responsible for bringing up topology. Will be instantiated using no-arg
